# edwards: optimize point negation