--- a/ecc/bls12-381/bandersnatch/point.go
+++ b/ecc/bls12-381/bandersnatch/point.go
@@ -181,8 +181,8 @@ func (p *PointAffine) IsOnCurve() bool {
 
 // Neg sets p to -p1 and returns it
 func (p *PointAffine) Neg(p1 *PointAffine) *PointAffine {
-	p.Set(p1)
-	p.X.Neg(&p.X)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
 	return p
 }
 
@@ -312,8 +312,9 @@ func (p *PointProj) IsZero() bool {
 // Neg negates point (x,y) on a twisted Edwards curve with parameters a, d
 // modifies p
 func (p *PointProj) Neg(p1 *PointProj) *PointProj {
-	p.Set(p1)
-	p.X.Neg(&p.X)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
+	p.Z = p1.Z
 	return p
 }
 
@@ -445,9 +446,10 @@ func (p *PointExtended) Equal(p1 *PointExtended) bool {
 // Neg negates point (x,y) on a twisted Edwards curve with parameters a, d
 // modifies p
 func (p *PointExtended) Neg(p1 *PointExtended) *PointExtended {
-	p.Set(p1)
-	p.X.Neg(&p.X)
-	p.T.Neg(&p.T)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
+	p.Z = p1.Z
+	p.T.Neg(&p1.T)
 	return p
 }