--- a/internal/generator/edwards/template/point.go.tmpl
+++ b/internal/generator/edwards/template/point.go.tmpl
@@ -166,8 +166,8 @@ func (p *PointAffine) IsOnCurve() bool {
 
 // Neg sets p to -p1 and returns it
 func (p *PointAffine) Neg(p1 *PointAffine) *PointAffine {
-	p.Set(p1)
-	p.X.Neg(&p.X)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
 	return p
 }
 
@@ -297,8 +297,9 @@ func (p *PointProj) IsZero() bool {
 // Neg negates point (x,y) on a twisted Edwards curve with parameters a, d
 // modifies p
 func (p *PointProj) Neg(p1 *PointProj) *PointProj {
-	p.Set(p1)
-	p.X.Neg(&p.X)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
+	p.Z = p1.Z
 	return p
 }
 
@@ -458,9 +459,10 @@ func (p *PointExtended) Equal(p1 *PointExtended) bool {
 // Neg negates point (x,y) on a twisted Edwards curve with parameters a, d
 // modifies p
 func (p *PointExtended) Neg(p1 *PointExtended) *PointExtended {
-	p.Set(p1)
-	p.X.Neg(&p.X)
-	p.T.Neg(&p.T)
+	p.X.Neg(&p1.X)
+	p.Y = p1.Y
+	p.Z = p1.Z
+	p.T.Neg(&p1.T)
 	return p
 }